move remaining `doc` constants to `rasa.shared`
