+++ rasa/cli/x.py
@@ -24,8 +24,8 @@ from rasa.constants import (
     DEFAULT_LOG_LEVEL_RASA_X,
     DEFAULT_RASA_X_PORT,
     DEFAULT_RASA_PORT,
-    DOCS_BASE_URL_RASA_X,
 )
+from rasa.shared.constants import DOCS_BASE_URL_RASA_X
 from rasa.core.utils import AvailableEndpoints
 from rasa.utils.endpoints import EndpointConfig
