+++ rasa/constants.py
@@ -1,7 +1,5 @@
 import os
 
-from rasa.shared.constants import DOCS_BASE_URL
-
 DEFAULT_ENDPOINTS_PATH = "endpoints.yml"
 DEFAULT_CREDENTIALS_PATH = "credentials.yml"
 DEFAULT_CONFIG_PATH = "config.yml"
@@ -28,17 +26,6 @@ PACKAGE_NAME = "rasa"
 
 DEFAULT_RASA_X_PORT = 5002
 DEFAULT_RASA_PORT = 5005
-
-DOCS_URL_TEST_STORIES = DOCS_BASE_URL + "/testing-your-assistant"
-DOCS_URL_ACTIONS = DOCS_BASE_URL + "/core/actions/"
-DOCS_URL_CONNECTORS = DOCS_BASE_URL + "/user-guide/connectors/"
-DOCS_URL_EVENT_BROKERS = DOCS_BASE_URL + "/api/event-brokers/"
-DOCS_URL_PIKA_EVENT_BROKER = DOCS_URL_EVENT_BROKERS + "#pika-event-broker"
-DOCS_URL_TRACKER_STORES = DOCS_BASE_URL + "/api/tracker-stores/"
-DOCS_URL_COMPONENTS = DOCS_BASE_URL + "/nlu/components/"
-DOCS_URL_MIGRATION_GUIDE = DOCS_BASE_URL + "/migration-guide/"
-
-DOCS_BASE_URL_RASA_X = "https://rasa.com/docs/rasa-x"
 
 MINIMUM_COMPATIBLE_VERSION = "2.0.0a3"
