+++ rasa/core/policies/form_policy.py
@@ -3,7 +3,7 @@ from typing import List, Dict, Text, Optional, Any, Union, Tuple
 
 import rasa.shared.utils.common
 import rasa.shared.utils.io
-from rasa.constants import DOCS_URL_MIGRATION_GUIDE
+from rasa.shared.constants import DOCS_URL_MIGRATION_GUIDE
 from rasa.shared.core.constants import (
     ACTION_LISTEN_NAME,
     LOOP_NAME,
@@ -165,7 +165,4 @@ class FormPolicy(MemoizationPolicy):
         return result
 
     def _metadata(self) -> Dict[Text, Any]:
-        return {
-            "priority": self.priority,
-            "lookup": self.lookup,
-        }
+        return {"priority": self.priority, "lookup": self.lookup}
