+++ rasa/shared/constants.py
@@ -7,6 +7,15 @@ DOCS_URL_RULES = DOCS_BASE_URL + "/core/rules/"
 DOCS_URL_FORMS = DOCS_BASE_URL + "/core/forms/"
 DOCS_URL_PIPELINE = DOCS_BASE_URL + "/nlu/choosing-a-pipeline/"
 DOCS_URL_POLICIES = DOCS_BASE_URL + "/core/policies/"
+DOCS_URL_TEST_STORIES = DOCS_BASE_URL + "/testing-your-assistant"
+DOCS_URL_ACTIONS = DOCS_BASE_URL + "/core/actions/"
+DOCS_URL_CONNECTORS = DOCS_BASE_URL + "/user-guide/connectors/"
+DOCS_URL_EVENT_BROKERS = DOCS_BASE_URL + "/api/event-brokers/"
+DOCS_URL_PIKA_EVENT_BROKER = DOCS_URL_EVENT_BROKERS + "#pika-event-broker"
+DOCS_URL_TRACKER_STORES = DOCS_BASE_URL + "/api/tracker-stores/"
+DOCS_URL_COMPONENTS = DOCS_BASE_URL + "/nlu/components/"
+DOCS_URL_MIGRATION_GUIDE = DOCS_BASE_URL + "/migration-guide/"
+DOCS_BASE_URL_RASA_X = "https://rasa.com/docs/rasa-x"
 
 INTENT_MESSAGE_PREFIX = "/"
